+++ src/freenet/node/PeerMessageQueue.java
@@ -1,5 +1,7 @@
 package freenet.node;
 
+import static java.util.concurrent.TimeUnit.MINUTES;
+
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -42,13 +44,13 @@ public class PeerMessageQueue {
 	private class PrioQueue {
 		
 		// FIXME refactor into PrioQueue and RoundRobinByUIDPrioQueue
-		PrioQueue(int timeout, boolean timeoutSinceLastSend) {
+		PrioQueue(long timeout, boolean timeoutSinceLastSend) {
 			this.timeout = timeout;
 			this.roundRobinBetweenUIDs = timeoutSinceLastSend;
 		}
 		
 		/** The timeout, period after which messages become urgent. */
-		final int timeout;
+		final long timeout;
 		/** If true, do round-robin between UID's, and count the timeout relative
 		 * to the last send. Block transfers need this - both realtime and bulk. */
 		final boolean roundRobinBetweenUIDs;
